Image_segmentation/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
matplotlib
streamlit